Update README.md
Overview:
·         1. Our Python application connects successfully over server‑only SSL using a .kdb file and username/password.
·         2. Our .NET client, built with the IBM XMS managed code library, consistently returns MQRC_SSL_INITIALIZATION_ERROR (2393).
We understand that, unlike Python or Java, the .NET managed client relies on the operating systems certificate store. Since our application is containerized, running on a Linux distro, we have tried two separate mechanisms to register the certificate:
To the user certificates in the deployment script
 
To the user certificates via .NET code.
 
Via both method 1 & 2 we can see the certs are loaded in the keystore via .NET (localmachine for method 1, user for method2):
 
We have also tried setting some of the extended SSL properties as below with no luck:
 
Any guidance you could offer would be greatly appreciated.